# Avoid mixing SSE and AVX in XTS-mode AVX512 implementation